moved atom construction to utils and added debug mode
